feature/migrate to spring boot v2.7.18
