--- src/enums/BookGenre.java
@@ -1,20 +1,0 @@
-package enums;
-
-public enum BookGenre {
-    FICTION,
-    NON_FICTION,
-    MYSTERY,
-    SCIENCE_FICTION,
-    FANTASY,
-    BIOGRAPHY,
-    HISTORY,
-    ROMANCE,
-    THRILLER;
-
-    @Override
-    public String toString(){
-        return name().replace("_", " ").toLowerCase();
-    }
-}
-
-
