--- src/enums/ReservationStatus.java
@@ -1,8 +1,0 @@
-package enums;
-
-public enum ReservationStatus {
-    PENDING,
-    APPROVED,
-    REJECTED
-}
-
