--- src/enums/UserRole.java
@@ -1,8 +1,0 @@
-package enums;
-
-public enum UserRole {
-    ADMIN,
-    MANAGER,
-    REGULAR_USER,
-    USER
-}
