+++ src/main/java/com/example/Librarymanagementsystem/LibraryManagementSystemApplication.java
@@ -1,0 +1,13 @@
+package com.example.Librarymanagementsystem;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LibraryManagementSystemApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LibraryManagementSystemApplication.class, args);
+	}
+
+}
