+++ src/main/java/com/example/Librarymanagementsystem/config/SecurityConfig.java
@@ -1,0 +1,4 @@
+package com.example.Librarymanagementsystem.config;
+
+public class SecurityConfig {
+}
