+++ src/main/java/com/example/Librarymanagementsystem/exception/AppException.java
@@ -1,0 +1,4 @@
+package com.example.Librarymanagementsystem.exception;
+
+public class AppException {
+}
