+++ src/main/java/com/example/Librarymanagementsystem/model/enums/BookGenre.java
@@ -1,0 +1,20 @@
+package com.example.Librarymanagementsystem.model.enums;
+
+public enum BookGenre {
+    FICTION,
+    NON_FICTION,
+    MYSTERY,
+    SCIENCE_FICTION,
+    FANTASY,
+    BIOGRAPHY,
+    HISTORY,
+    ROMANCE,
+    THRILLER;
+
+    @Override
+    public String toString(){
+        return name().replace("_", " ").toLowerCase();
+    }
+}
+
+
