+++ src/main/java/com/example/Librarymanagementsystem/model/enums/ReservationStatus.java
@@ -1,0 +1,8 @@
+package com.example.Librarymanagementsystem.model.enums;
+
+public enum ReservationStatus {
+    PENDING,
+    APPROVED,
+    REJECTED
+}
+
