+++ src/main/java/com/example/Librarymanagementsystem/model/enums/UserRole.java
@@ -1,0 +1,8 @@
+package com.example.Librarymanagementsystem.model.enums;
+
+public enum UserRole {
+    ADMIN,
+    MANAGER,
+    REGULAR_USER,
+    USER
+}
