+++ src/main/java/com/example/Librarymanagementsystem/payload/request/SignUpRequest.java
@@ -1,0 +1,4 @@
+package com.example.Librarymanagementsystem.payload.request;
+
+public class SignUpRequest {
+}
