+++ src/main/java/com/example/Librarymanagementsystem/payload/response/SignUpResponse.java
@@ -1,0 +1,4 @@
+package com.example.Librarymanagementsystem.payload.response;
+
+public class SignUpResponse {
+}
