+++ src/main/java/com/example/Librarymanagementsystem/repository/UserRepository.java
@@ -1,0 +1,5 @@
+package com.example.Librarymanagementsystem.repository;
+
+public class UserRepository{
+
+}
