+++ src/main/java/com/example/Librarymanagementsystem/service/BookService.java
@@ -1,0 +1,4 @@
+package com.example.Librarymanagementsystem.service;
+
+public class BookService {
+}
