--- src/model/Utils.java
@@ -1,30 +1,0 @@
-package model;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-public class Utils {
-
-    public static String numberToHash(String password) {
-        try {
-
-            MessageDigest digest = MessageDigest.getInstance("SHA-256");
-
-            byte[] hashBytes = digest.digest(password.getBytes());
-
-            StringBuilder hexString = new StringBuilder();
-            for (byte b : hashBytes) {
-                String hex = Integer.toHexString(0xff & b);
-                if (hex.length() == 1) {
-                    hexString.append('0');
-                }
-                hexString.append(hex);
-            }
-
-            return hexString.toString();
-
-        } catch (NoSuchAlgorithmException e) {
-            throw new RuntimeException("No such hashing algorithm", e);
-        }
-    }
-}
